edit

+++ tasks/viewsets.py
@@ -1,13 +1,12 @@
 from rest_framework import viewsets
 from rest_framework.parsers import JSONParser
-
-from . import models
-from . import serializers
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from tasks.models import Tasks
 from tasks.serializers import TasksSerializer
+from rest_framework.views import APIView
+from rest_framework.permissions import IsAdminUser
 
 
 @api_view(['GET', 'POST'])
@@ -40,8 +39,14 @@ def task_view(request, id):
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
-        return Response(serializer.errors, status= status.HTTP_404_BAD_REQUEST)
+        return Response(serializer.errors, status= 404)
 
     elif request.method == 'DELETE':
-        task.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        # usually I will have a session information from login, but since I don't have that I just wrote how I would normally execute authentification
+        if request.user.type != 10:
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            task.delete()
+            return Response(status=200)
+
+
